requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
beautifulsoup4==4.5.1     # via wagtail
billiard==3.3.0.23        # via celery
boto==2.39.0
celery==3.1.23
contextlib2==0.5.4        # via raven
defusedxml==0.4.1         # via python3-openid
dj-database-url==0.3.0
dj-static==0.0.6
django-discover-runner==1.0  # via django-role-permissions
django-modelcluster==2.0  # via wagtail
django-role-permissions==1.2
django-server-status==0.3.2
django-storages-redux==1.3.2
django-taggit==0.18.3     # via wagtail
django-treebeard==4.0.1   # via wagtail
django-webpack-loader==0.4.1
django==1.10.3
djangorestframework==3.5.2
edx-api-client==0.3.0
edx-opaque-keys==0.4
elasticsearch-dsl==2.1.0
elasticsearch==2.3.0
factory-boy==2.7.0
fake-factory==0.7.2       # via factory-boy
faker==0.7.3
html5lib==0.999999
jsonfield==1.0.3
kombu==3.0.37             # via celery, django-server-status
newrelic==2.72.1.53
oauthlib==2.0.0           # via python-social-auth, requests-oauthlib
pbr==1.10.0               # via stevedore
pillow==3.4.2
psycopg2==2.6
pyjwt==1.4.2              # via python-social-auth
pymongo==3.4.0            # via edx-opaque-keys
python-dateutil==2.5.3
python-social-auth==0.2.21
python3-openid==3.0.10    # via python-social-auth
pytz==2016.7              # via celery, django-modelcluster
PyYAML==3.11
raven==5.32.0
redis==2.10.5
requests-oauthlib==0.7.0  # via python-social-auth
requests==2.11.1
six==1.10.0               # via django-role-permissions, django-server-status, edx-api-client, edx-opaque-keys, elasticsearch-dsl, fake-factory, faker, html5lib, python-dateutil, python-social-auth, stevedore
static3==0.5.1
stevedore==1.19.0         # via edx-opaque-keys
unidecode==0.4.19         # via wagtail
urllib3==1.19             # via elasticsearch
uwsgi==2.0.14
wagtail==1.7
willow==0.4               # via wagtail
